+++ CHANGELOG.rst
@@ -10,6 +10,14 @@ Changelog
 * API Change: clarified naming. renamed "boundary" to "bbox". renamed "polygon" to "boundary". boundaries (the outer polygon defining part of a timezone) and holes are both polygons so hence the name "polygon" is ambiguous.
 
 
+6.6.1 (2025-07-18)
+------------------
+
+* hotfix missing `flatbuf` module in the distributions
+* added tests for checking the content of the distributions
+
+
+
 6.6.0 (2025-07-17)
 ------------------